refactor: #306 - HomeViewReactor+Extension에 extensions 분리
